% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-rollup.R
\name{byRollup}
\alias{byRollup}
\title{Aggregate over a log data.frame by a unit of time}
\usage{
byRollup(logdf, unit = c("second", "minute", "hour", "day", "week", "month",
  "year", "quarter"), FUN = elbr:::summarizeLog)
}
\arguments{
\item{logdf}{data.frame as returned by \code{\link{cleanLog}}. Or rather,
a data.frame with a "timestamp" variable that is a datetime class.}

\item{unit}{character unit of time, passed to
\code{\link[lubridate]{floor_date}}.}

\item{FUN}{function to apply to each chunk of \code{logdf}. Default is
\code{\link{summarizeLog}}.}
}
\value{
A data.frame with the time values as the rownames.
}
\description{
Aggregate over a log data.frame by a unit of time
}

