% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ELBLog}
\alias{ELBLog}
\title{Interface to a set of ELB log files}
\usage{
ELBLog(start_date = NULL, end_date = start_date,
  path = getOption("elbr.dir", "."))
}
\arguments{
\item{start_date}{A \code{Date} or character in format "YYYY-MM-DD" representing
a starting date for log files to consider. Default is \code{NULL}, meaning that
all log files found in \code{path} will be considered. Specifying a \code{start_date}
when appropriate can greatly speed up analyses by skipping files that are
out of the range of interest.}

\item{end_date}{See \code{start_date}. Default for \code{end_date} is the value of
\code{start_date}, i.e. if a \code{start_date} is given and no \code{end_date} is provided,
only that day's log files will be considered.}

\item{path}{character file path indicating where the ELB logs are contained.
Default is the current working directory, unless \code{options(elbr.dir)} is set.}
}
\value{
An S3-class 'ELBLog' object.
}
\description{
AWS ELB logs are stored in a series of paths separated for each year, then
month, then day, with many files per day in the lowest level directories.
Objects of class \code{ELBLog} contain information about which log files to read
in, and methods such as \code{\link[=select]{select()}}, \code{\link[=filter]{filter()}}, \code{\link[=summarize]{summarize()}}, and
\code{\link[=collect]{collect()}} are defined to enable you to work with this log data as if it
were a single \code{data.frame} in memory, even though constructing that dataset
would be too slow or perhaps impossible due to resource constraints.
}
