% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-logs.R
\name{filterLogs}
\alias{filterLogs}
\alias{find504s}
\title{Grab log entries that match a pattern}
\usage{
filterLogs(pattern = "", date = "", base.dir = getOption("elbr.dir"))

find504s(date = "", base.dir = getOption("elbr.dir"))
}
\arguments{
\item{pattern}{character regular expression pattern to pass to 'egrep'}

\item{date}{character date string of the format YYYY/MM/DD, or any segment of
that (e.g. YYYY/MM).}

\item{base.dir}{character directory in which the ELB logs are located}
}
\value{
A data.frame, via 'read.elb'
}
\description{
'filterLogs' can take any pattern you give it. 'find504s' calls 'filterLogs'
with the pattern that identifies 504 Gateway Timeout responses.
}
