% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-logs.R
\name{find504s}
\alias{find504s}
\title{Grab log entries that identifie 504 Gateway Timeouts}
\usage{
find504s(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=mapELB]{mapELB()}}}
}
\value{
A data.frame (tibble), via 'read_elb'
}
\description{
Grab log entries that identifie 504 Gateway Timeouts
}
