% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elb.R
\name{parse_request}
\alias{parse_request}
\title{Split the 'request' into verb, url, protocol}
\usage{
parse_request(x)
}
\arguments{
\item{x}{Character vector, the "request" column from an ELB data.frame}
}
\value{
A \code{data.frame} with three columns: "request_verb", "request_url",
and "request_protocol".
}
\description{
Split the 'request' into verb, url, protocol
}
