% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elb.R
\name{read.elb}
\alias{read.elb}
\title{Load an ELB log file}
\usage{
read.elb(file, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{A file name or connection. See \code{\link[utils]{read.table}}}

\item{stringsAsFactors}{The standard \code{data.frame} argument, but
defaulted to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{read.delim}}
}
\value{
A \code{data.frame}.
}
\description{
Load an ELB log file
}

