% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{standardizeURLs}
\alias{standardizeURLs}
\title{Categorize URLs by removing specifics}
\usage{
standardizeURLs(url)
}
\arguments{
\item{url}{character vector of request URLs}
}
\value{
A character vector of equal length containing URLs with the entity
IDs and query strings removed.
}
\description{
Abstract away from entity ids and query specifics to see if certain endpoints
generally behave a certain way.
}
\examples{
standardizeURLs("https://crunch.io/api/datasets/000001/") == standardizeURLs("https://crunch.io/api/datasets/999999/")
}
