% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{summarize504s}
\alias{summarize504s}
\title{Make a summary of Gateway Timeouts}
\usage{
summarize504s(days, before.date = Sys.Date(), send = TRUE)
}
\arguments{
\item{days}{Integer number of days to aggregate}

\item{before.date}{days before this day. Default is today (i.e. with days=1,
give the report for yesterday)}

\item{send}{Logical: send messages to Slack?}
}
\description{
Make a summary of Gateway Timeouts
}
