% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarizeLog}
\alias{summarizeLog}
\title{Basic summary statistics from a log}
\usage{
summarizeLog(logdf)
}
\arguments{
\item{logdf}{data.frame as returned by \code{\link{cleanLog}}.}
}
\value{
A data.frame with one row.
}
\description{
Basic summary statistics from a log
}

