% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyzeELB}
\alias{analyzeELB}
\alias{mapELB}
\alias{reduceELB}
\title{Perform some analysis across a bunch of ELB log files}
\usage{
analyzeELB(...)

mapELB(FUN, start_date = NULL, end_date = start_date,
  files = find_log_files(start_date, end_date), select_vars = TRUE)

reduceELB(results)
}
\arguments{
\item{...}{Arguments passed to \code{mapELB}}

\item{FUN}{A function to map over each log data frame}

\item{start_date}{\code{Date} or date string specifying a starting point for
finding log files. Use \code{NULL}, the default, to get all log files.}

\item{end_date}{\code{Date} or date string to close the range. Default is
\code{start_date}, i.e. it takes all log files for a single day, if one is given.}

\item{files}{Vector of file names to read in; an alternative to specifying
date ranges.}

\item{select_vars}{logical: should we only read in the variables named in
\code{FUN}? Default is \code{TRUE}, which provides a speed boost by not loading data
we don't need for the current analysis, but in cases where \code{FUN} doesn't
name every variable we want to keep, set this to \code{FALSE}.}

\item{results}{For \code{reduceELB}, the results of \code{mapELB}}
}
\value{
A list of query results.
}
\description{
Perform some analysis across a bunch of ELB log files
}
