% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elb.R
\name{cleanLog}
\alias{cleanLog}
\title{Do some general cleaning}
\usage{
cleanLog(logdf)
}
\arguments{
\item{logdf}{a \code{data.frame} as returned from \code{\link{read.elb}}}
}
\value{
A \code{data.frame} cleaned up a bit.
}
\description{
Delete some columns we don't care about ever, parse the timestamp,
add up response times, separate request verb from URL, etc.
}
